Refactored sequence generation
Included hold percentage and pattern controls.

+++ V3.0/ZMMT/src/nl/zeesoft/zmmt/sequencer/SeqNote.java
@@ -2,8 +2,8 @@ package nl.zeesoft.zmmt.sequencer;
 
 public class SeqNote {
 	public String	instrument	= "";
-	public int		tickStart	= 0;
-	public int		tickEnd		= 0;
+	public long		tickStart	= 0;
+	public long		tickEnd		= 0;
 	public int		midiNote	= 0;
 	public int		velocity	= 0;
 	public int		channel		= 0;
